requirements.txt
asyncio
logging
python-dotenv
load_dotenv
aiogram
sqlalchemy
aiosqlite
matplotlib